Move Api code to main Restid module
